Add testing code for testing parse_lines fxn
Having trouble getting the parse_lines function to work, with it returning a "Exception: No match in record selector line_num", which is really strange because we're attempting to create a Parsed_line object. At first this looked like a problem in the parse_lines function where we use the line_num record selector to get a line number from an Unparsed_line object, but more careful testing has the exception arising even when I don't explicitly use the line_num record selector in trying to build the Parsed_line object. Very puzzling.