Add parens to instance for show AddExpr
A trivial tweak to the instance def for show (AddExpr e1 2) to help fix a small oddity where expressions intended to represent something like 2 * (3 + 4) were printing as 2 * 3 + 4. This is just a change to the show output and doesn't appear to change the (already correct) internal representation that was parsing and evaluating ok.