Rearrange ordering in expr parser & value parser
A bit of rearrangement in the order of +++ terms in the expr parser and value parser helps deal with some parsing difficulties for expressions like INT(RND(5) * H + 1) and INT(RND(5) * (H + 1)) needed in the test.bas program. Having to play with such specific parsing order seems highly unsatisfactory in the long run , suggesting too much context-specificity in the types established in BasicTypes.hs, but this might be sufficient to move forward on the initial limited interpreter. Commit also includes some extra testing elements.