Add nextlist_statement parser for things like "NEXT A, B, C"
This is working but still needs some cleaning up and a little more testing. This additional parser is meant to deal with the more general NEXT statements that involve a comma-separated list if Vars, like "NEXT A, B, C". This commit then also includes some helper-parsers for the list-like construction "A, B, C". Realize that we will probably merge NEXTLIST into NEXT shortly once testing is complete.