Add expr_colon_tab parser to catch "PRINT TAB(#);" statements
Previous tab_print_expr parser refused to work (not catching the "TAB" string), so I eventually added this expr_colon_tab parser at a higher level which avoids the indirection of the expr parser. This seems to work fine now, catching the print statements of the form:
PRINT TAB(#); "Hello!"
and producing the # blank spaces wanted.