Reverse order of next Parsers in statement Parser list
We thought perhaps we hadn't yet created the nextlist_statement Parser, but it was already there. This commit reverses the order in which the next_statement and nextlist_statement appear in the more general statement Parser so that it checks for the nextlist_statement before it checks for teh next_statement. Can't check this process yet because of the broken print_expression fxn.